rbd: move snapshot test funcs to a new file
